refactor: reorganize store slices and update imports for chatbot and text animation
