+++ src/store/slices/chatbot/text-animation-slice.ts
